fix / adjustments

+++ src/controllers/urls.controllers.js
@@ -43,7 +43,7 @@ export async function getUrl (req, res){
 
 export async function getShortUrl (req, res){
     
-    const { shortUrl } = req.body;
+    const { shortUrl, visit } = req.params;
 
     try {
       const listShortUrl = await db.query(
@@ -51,14 +51,19 @@ export async function getShortUrl (req, res){
         [shortUrl]);
 
         if (listShortUrl.rowCount <= 0) return res.status(404).send("Not Found!");
-        
-      const redirectUrl = listShortUrl.rows[0].url;
-      const countVisit = listShortUrl.rows[0].visit + 1;
-  
-      await db.query(
-        `UPDATE urls visit SET visit = $1 WHERE "shortUrl"= $2`, 
-        [countVisit, shortUrl]);
 
+        const listVisit = await db.query(
+            `SELECT * FROM urls WHERE visit = $1`, 
+            [visit]);
+
+        const countVisit = listVisit.rows[0].visit + 1;
+
+        await db.query(
+            `UPDATE urls SET visit = $1 WHERE "shortUrl"=$2`, 
+            [countVisit, shortUrl]);
+        
+        const redirectUrl = listShortUrl.rows[0].url;
+        
       res.redirect(redirectUrl);
     } catch (error) {
       res.status(500).send(error.message);
@@ -68,26 +73,30 @@ export async function getShortUrl (req, res){
 
 export async function deleteUrl (req, res){
 
-    const {id} = req.params
-    const { userId } = req.locals.session;
+    const { id } = req.params.id;
+    const { shortUrl } = req.params;
 
-    try {
-        const urlExists  = await db.query(`SELECT * FROM urls WHERE id = $1`, 
-        [id]);
+  try {
+    const deleteShortUrl = await db.query(`SELECT * FROM urls WHERE id=$1`, [id]
+    );
 
-        if (urlExists.rowCount <= 0) return res.status(404).send("Not Found!");
-       
-        const userExists = await db.query(`SELECT * FROM urls WHERE id = $1, "userId" = $2`, 
-        [id, userId]);
+    if (deleteShortUrl.rowCount <= 0) return res.status(404).send("Not Found!");
 
-        if (userExists.rowCount <= 0) return res.status(401).send("ID Not Found!");
-  
-      await db.query(`DELETE FROM urls WHERE id=$1`,
-        [id]);
-      
-      return res.sendStatus(204);
+    const listShortUrl = await db.query(
+        `SELECT * FROM urls WHERE "shortUrl" = $1`, 
+        [shortUrl]);
+
+    if (deleteShortUrl.rows[0] != listShortUrl.rows[0]) return res.status(401).send("ID Not Found!")
+ 
+    if (deleteShortUrl.rows[0] === listShortUrl.rows[0]){
+        await db.query(`DELETE FROM urls WHERE id=$1`,
+      [id]);
     }
-    catch{
-      return res.status(500).send(err.message)
-    }  
-  };
+    
+    return res.sendStatus(204);
+  }
+  catch{
+    return res.status(500).send(err.message)
+  }
+
+}
